checking user and email exists

+++ api/localStorageApi.js
@@ -41,25 +41,45 @@ const registerUser = (login, email, password) => {
         password: password,
     }
 
-    setCurrentUser(newUser)
+    const isUserExist = checkIfUserExist(newUser, users)
+    const isEmailExists = checkIfEmailExists(newUser, users)
 
-    window.localStorage.setItem(
-        "users",
-        JSON.stringify(new Array(newUser, ...users))
-    )
+    if (!isUserExist && !isEmailExists) {
+        window.localStorage.setItem(
+            "users",
+            JSON.stringify(new Array(newUser, ...users))
+        )
+        changeUsername(newUser.login)
+        setCurrentUser(newUser)
+        navigation.login()
+    } else {
+        navigation.unsetHidden(modals[0])
+    }
+}
 
-    changeUsername(newUser.login)
-    navigation.login()
+const checkIfUserExist = (currentUser, users) => {
+    for (let user of users) {
+        if (currentUser.login === user.login) {
+            console.log("user exists")
+            navigation.unsetHidden(errors[6])
+            return 1
+        } else return 0
+    }
+}
+
+const checkIfEmailExists = (currentUser, users) => {
+    for (let user of users) {
+        if (currentUser.email === user.email) {
+            console.log("email exists")
+            navigation.unsetHidden(errors[7])
+            return 1
+        } else return 0
+    }
 }
 
 initUserApi = () => {
     document.addEventListener("DOMContentLoaded", () => {
         getUsers()
         getCurrentUser()
-
-        console.log(users)
-        console.log(currentUser)
     })
 }
-
-initUserApi()
